Upload Page Created
A new page where I will go and upload my projects has been created. I also created a form on that page that will allow me to assign a title, description, links to github and a website (if applicable), up to 3 screenshots of the project, and the project type (Modeling and SImulation, Web Development, or Game Development)

As of right now, the form doesn't do anything, but my next task will be to connect it to the database and upload the information to the table.