Uploading functions created
A separate php file, upload-project.php, has been created that will upload the information inputted in the form. This will take that information and insert it into the Projects table in the Portfolio database.

Bug:
While the destination of the images are uploaded to the database, the images are not seen in the actual uploads folder.